+++ src/main/java/lotto/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto;
 
 import java.util.HashSet;
 import java.util.List;
