+++ src/main/java/lotto/Score.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto;
 
 import java.util.Arrays;
 import java.util.List;
