+++ src/test/java/lotto/LottoServiceTest.java
@@ -3,11 +3,6 @@ package lotto;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import lotto.domain.Lotto;
-import lotto.domain.Score;
-import lotto.service.LottoGenerator;
-import lotto.service.LottoResultChecker;
-import lotto.service.LottoService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
